requirements.txt
pyTelegramBotAPI==4.14.0
openai-whisper==20231117
python-dotenv==1.0.0
